make doc more readable
Some subheads are added, since each docstring is long and hard to
distinguish. Currently they are defined manually, and I am not sure if
it is possible to use something like `autodoc` to achieve this part.

We should be more focused on Tutorial and Quick Start in future. They
are the parts that all new ones first visit.